+++ src/app/mod.rs
@@ -199,6 +199,7 @@ impl Liwan {
                     id: project_id.to_string(),
                     display_name: display_name.to_string(),
                     public,
+                    unlisted: false,
                     secret: None,
                 },
                 &[],